requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
alembic==1.18.4
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
auto-py-to-exe==2.50.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
colorlog==6.10.1
config==0.5.1
Eel==0.18.2
frozenlist==1.8.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
importlib_resources==6.5.2
input-with-timeout==2.2
Jinja2==3.1.6
keyboard==0.13.5
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.19.0
multidict==6.7.1
numpy==2.4.3
nvidia-ml-py==13.595.45
optuna==4.7.0
packaging==26.0
pandas==3.0.1
pefile==2024.8.26
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pynvml==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.6.3
win32_setctime==1.2.0
WMI==1.5.1
wsproto==1.2.0
yarl==1.22.0
zope.event==6.1
zope.interface==8.2
